microEye v2.3.2
New Features:

- Dynamic version management:
	- Implemented dynamic versioning and requirements handling in
      `_version.py` and `setup.py` instead of hardcoded values.

- PycroManager Integration:
	- Added support for managing multiple `PycroCore` instances to
      enable simultaneous bridge connections.
	- Implemented `HeadlessInstance` management through
      `HeadlessManager` for running micro-manager instances in the
      background (JavaBackend Only).
	- Created serialization/deserialization of Pycromanager instances in
      the configuration system.
	- Implemented `PycroCamera` and `PycroPanel` for camera acquisition.
	- Added `BridgesWidget` to manage `PycroCore` instances.
	- Added `HeadlessManagerWidget` to start/stop micro-manager in
      headless mode (JavaBackend Only).
	- Integrated `HeadlessOptions` widget for selecting headless
      operation parameters.

- Device Manager:
	- Migrated device management methods from `miEye_module` to
      `DeviceManager`, including adding/removing devices,
      configuration handling, and auto-connect/shutdown sequences.
	- Implemented comprehensive get_config/load_config system with
      hierarchical structure for complete system state persistence.
	- Centralized device functionality to facilitate future modular
      automation and integration of new hardware (work in progress).

- Camera Panel Shortcuts:
	- Added context menu for camera panels to facilitate changing
      frequently used properties and dynamic experiment naming
      options.

- Demo Cameras:
	- added `DemoLineScanner` to simulate linear detector array
      functionality for IR focus stabilization systems.
	- Implemented Z-Drift simulation, available under the Gaussian beam
      mode for demo cameras and arrays.

- miEye Module:
	- Migrated device-related methods into `DeviceManager` and removed
      deprecated code.
	- Implemented generic methods to handle added/removed device widgets
      instead of hardcoded solutions.
	- Improved dock widget save/load configuration methods.

- `KinesisXY`:
	- Refactored `runAsync` into `run_async` and moved it from View into
      the Controller.
	- Fixed issue where X and Y commands were being sent synchronously
      instead of asynchronously.

- `PzFoc`:
	- Added adjustable maximum range, particularly valid for FOCX00
      where previously FOC100 was assumed.
	- Refactored controller logic from the view.

- `FocusStabilizer`:
	- Fixed worker function bug where linear array detectors incorrectly
      used 2D image linear-ROI instead of data.
	- Added `toggleFocusStabilization` method to reflect stabilization
      status in the GUI.

- Various minor bug fixes and formatting adjustments.

Future Updates:
	- Finalized Pycro-Manager Integration with Z-Stage and Focus
      Stabilization.
	- MacOS compatibility testing
	- Acquisition Experiments Designer tools (continued development)
	- 3D Localization fitting from Experimental PSF
	- Introduction of 2D/3D Single-Particle Tracking

+++ src/microEye/hardware/widgets/controller.py
@@ -9,10 +9,10 @@ from microEye.utils.gui_helper import GaussianOffSet
 
 
 class Controller(QtWidgets.QDockWidget):
-    stage_move_requested = Signal(str, int, int)  # axis, direction, mode
-    stage_stop_requested = Signal(str)  # axis
-    stage_home_requested = Signal(str)  # axis
-    stage_toggle_lock = Signal(str)  # axis
+    stage_move_requested = Signal(str, int, float)
+    stage_stop_requested = Signal(str)
+    stage_home_requested = Signal(str)
+    stage_toggle_lock = Signal(str)
 
     def __init__(self):
         ''' '''
